set up assignment fetch for instructor and added button to add assignment
