+++ client/src/components/App/NavMenu/Calendar.jsx
@@ -4,39 +4,6 @@ import { format } from "date-fns"; // Using date-fns for date formatting
 const Calendar = () => {
   const [selectedDate, setSelectedDate] = useState(new Date());
   const [selectedDay, setSelectedDay] = useState(null);
-  const [assignmentDetails, setAssignmentDetails] = useState({});
-  const [showAssignmentInput, setShowAssignmentInput] = useState(false);
-
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [newAssignment, setNewAssignment] = useState({
-    title: "",
-    description: "",
-    dueDate: null,
-  });
-
-  const openModal = () => {
-    setIsModalOpen(true);
-  };
-
-  const closeModal = () => {
-    setIsModalOpen(false);
-  };
-
-  const handleInputChange = (event) => {
-    const { name, value } = event.target;
-    setNewAssignment((prevAssignment) => ({
-      ...prevAssignment,
-      [name]: value,
-    }));
-  };
-
-  const handleAssignmentSubmit = (event) => {
-    event.preventDefault();
-    if (newAssignment.title && newAssignment.dueDate) {
-      setAssignmentDetails(newAssignment);
-      closeModal();
-    }
-  };
 
   const handlePrevMonth = () => {
     setSelectedDate(
@@ -129,88 +96,6 @@ const Calendar = () => {
         <button onClick={handleNextMonth}>Next</button>
       </div>
       {calendarGrid}
-      <button
-        className="mt-2 p-2 text-1xl text-white bg-[#31503b] rounded-[10rem]"
-        onClick={openModal}
-      >
-        Add Assignment
-      </button>
-      {isModalOpen && (
-        <div className="fixed inset-0 flex items-center justify-center modal-overlay">
-          <div className="w-full p-4 bg-white rounded shadow-lg modal sm:w-96">
-            <h2 className="mb-4 text-xl font-semibold text-black">
-              Add Assignment
-            </h2>
-            <form onSubmit={handleAssignmentSubmit}>
-              <div className="mb-4">
-                <label
-                  htmlFor="title"
-                  className="block mb-1 font-medium text-black"
-                >
-                  Assignment Title:
-                </label>
-                <input
-                  type="text"
-                  id="title"
-                  name="title"
-                  value={newAssignment.title}
-                  onChange={handleInputChange}
-                  className="w-full p-2 text-black border border-gray-300 rounded"
-                  required
-                />
-              </div>
-              <div className="mb-4">
-                <label
-                  htmlFor="description"
-                  className="block mb-1 font-medium text-black"
-                >
-                  Description:
-                </label>
-                <textarea
-                  id="description"
-                  name="description"
-                  value={newAssignment.description}
-                  onChange={handleInputChange}
-                  className="w-full p-2 text-black border border-gray-300 rounded"
-                  required
-                />
-              </div>
-              <div className="mb-4">
-                <label
-                  htmlFor="dueDate"
-                  className="block mb-1 font-medium text-black"
-                >
-                  Due Date:
-                </label>
-                <input
-                  type="date"
-                  id="dueDate"
-                  name="dueDate"
-                  value={newAssignment.dueDate}
-                  onChange={handleInputChange}
-                  className="w-full p-2 text-black border border-gray-300 rounded"
-                  required
-                />
-              </div>
-              <div className="flex justify-end">
-                <button
-                  type="submit"
-                  className="px-4 py-2 mr-2 text-white bg-[#31503b] rounded"
-                >
-                  Submit Assignment
-                </button>
-                <button
-                  type="button"
-                  onClick={closeModal}
-                  className="px-4 py-2 text-gray-700 bg-gray-300 rounded"
-                >
-                  Cancel
-                </button>
-              </div>
-            </form>
-          </div>
-        </div>
-      )}
     </div>
   );
 };
